Add files via upload
The sms REST API connects to the student database created in MySQL. This API makes use of Spring-Boot-Data-Rest dependency, which in turn uses the JPA Repository to provide the fundamental CRUD operations on the database. The communication is done using JSON over http methods. Also provides security using the Spring-Security and Spring-Oauth-Client dependencies. The server port, database and Oauth2 client is configured using the application.properties file.